+++ src/random/M_Level.cs
@@ -2,12 +2,15 @@ using HarmonyLib;
 
 namespace IShowSeed.Random;
 
-[TogglablePatch]
+[PermanentPatch]
 [HarmonyPatch(typeof(M_Level), "Awake")]
 public static class M_Level_Awake_Patcher
 {
     public static void Prefix(M_Level __instance)
     {
-        __instance.canFlip = false;
+        if (Plugin.IsSeededRun())
+        {
+            __instance.canFlip = false;
+        }
     }
 }