+++ src/random/OS_Computer_Interface.cs
@@ -4,7 +4,7 @@ namespace IShowSeed.Random;
 
 
 // seed will be set by the mod, not by the game
-[TogglablePatch]
+[OnlyForSeededRunsPatch]
 [HarmonyPatch(typeof(OS_Computer_Interface), "SetSeed")]
 public static class OS_Computer_Interface_SetSeed_Patcher
 {