+++ src/ui/StatsMenu.cs
@@ -11,7 +11,7 @@ public static class UT_SeededEnable_OnEnable_Patcher
     {
         TextMeshProUGUI text = __instance.gameObject.GetComponent<TextMeshProUGUI>();
         text.text = $"SEED: {Plugin.SeedForRandom}";
-        if (Plugin.ConfigPresetSeed.Value != 0) {
+        if (Plugin.IsSeededRun()) {
             text.text = "PRESET " + text.text;
         }
         return false; // skips OnEnable which calls OnCheckStatus to hide the text when WorldLoader.customSeed = false